feat(code-editor): move testing utilities for CodeMirror extensions
