+++ packages/code-editor/src/CodeEditor/hooks/extensions/useHyperLinkExtension.spec.ts
@@ -3,7 +3,7 @@ import { renderHook } from '@testing-library/react';
 import {
   createMockHyperLinkModule,
   createMockStateModule,
-} from '../../../testing';
+} from '../hooks.testUtils';
 
 import { useHyperLinkExtension } from './useHyperLinkExtension';
