+++ packages/code-editor/src/testing/index.ts
@@ -1,19 +1,1 @@
-import { codeSnippets } from './codeSnippets';
-
-export { codeSnippets };
-
-// Extension testing utilities
-export {
-  createComprehensiveFakeModules,
-  createDefaultTestProps,
-  createLanguageModuleFactory,
-  createMockAutoCompleteModule,
-  createMockExtension,
-  createMockHyperLinkModule,
-  createMockLanguageModule,
-  createMockLezerHighlightModule,
-  createMockLintModule,
-  createMockStateModule,
-  createMockViewModule,
-  FakeCompartment,
-} from './extensionTestUtils';
+export { codeSnippets } from './codeSnippets';
